+++ packages/ai/src/ui/default-chat-store.ts
@@ -69,11 +69,6 @@ export interface DefaultChatStoreOptions<
   body?: object;
 
   /**
-    Streaming protocol that is used. Defaults to `ui-message`.
-       */
-  streamProtocol?: 'ui-message' | 'text';
-
-  /**
     Custom fetch implementation. You can use it as a middleware to intercept requests,
     or to provide a custom fetch implementation for e.g. testing.
         */
@@ -144,7 +139,6 @@ export function defaultChatStore<
     >({
       api,
       fetch,
-      streamProtocol,
       credentials,
       headers,
       body,