+++ packages/ai/src/ui/ui-messages.ts
@@ -59,7 +59,7 @@ export type UIMessagePart<DATA_TYPES extends UIDataTypes> =
   | TextUIPart
   | ReasoningUIPart
   | ToolInvocationUIPart
-  | SourceUIPart
+  | SourceUrlUIPart
   | FileUIPart
   | DataUIPart<DATA_TYPES>
   | StepStartUIPart;
@@ -116,19 +116,12 @@ export type ToolInvocationUIPart = {
 /**
  * A source part of a message.
  */
-export type SourceUIPart = {
-  type: 'source';
-
-  /**
-   * The source.
-   */
-  source: {
-    sourceType: 'url';
-    id: string;
-    url: string;
-    title?: string;
-    providerMetadata?: Record<string, any>;
-  };
+export type SourceUrlUIPart = {
+  type: 'source-url';
+  sourceId: string;
+  url: string;
+  title?: string;
+  providerMetadata?: Record<string, any>;
 };
 
 /**