+++ packages/google-vertex/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @ai-sdk/google-vertex
 
+## 3.0.0-alpha.3
+
+### Patch Changes
+
+- Updated dependencies [6b98118]
+  - @ai-sdk/provider@2.0.0-alpha.3
+  - @ai-sdk/anthropic@2.0.0-alpha.3
+  - @ai-sdk/google@2.0.0-alpha.3
+  - @ai-sdk/provider-utils@3.0.0-alpha.3
+
+## 3.0.0-alpha.2
+
+### Patch Changes
+
+- Updated dependencies [26535e0]
+  - @ai-sdk/provider@2.0.0-alpha.2
+  - @ai-sdk/anthropic@2.0.0-alpha.2
+  - @ai-sdk/google@2.0.0-alpha.2
+  - @ai-sdk/provider-utils@3.0.0-alpha.2
+
 ## 3.0.0-alpha.1
 
 ### Patch Changes