+++ packages/svelte/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @ai-sdk/svelte
 
+## 3.0.0-alpha.3
+
+### Patch Changes
+
+- Updated dependencies [ab7ccef]
+- Updated dependencies [257224b]
+- Updated dependencies [0463011]
+- Updated dependencies [d306260]
+  - ai@5.0.0-alpha.3
+  - @ai-sdk/provider-utils@3.0.0-alpha.3
+
+## 3.0.0-alpha.2
+
+### Patch Changes
+
+- Updated dependencies [82aa95d]
+  - ai@5.0.0-alpha.2
+  - @ai-sdk/provider-utils@3.0.0-alpha.2
+
 ## 3.0.0-alpha.1
 
 ### Patch Changes